+++ LiteCore/Query/QueryParser.cc
@@ -308,7 +308,7 @@ namespace litecore {
             str.insert((string::size_type)startPosOfWhat, extra.str());
             _sql.str(str);
             _sql.seekp(0, stringstream::end);
-            _1stCustomResultCol += 1 + _ftsTables.size();
+            _1stCustomResultCol += 1U + narrow_cast<unsigned int>(_ftsTables.size());
         }
 
         // ORDER_BY clause:
